+++ CHANGELOG.md
@@ -56,6 +56,8 @@ This project adheres to [Semantic Versioning](https://semver.org/).
   MacOS, FreeBSD, DragonflyBSD, Android, iOS and Haiku.
   ([#2074](https://github.com/nix-rust/nix/pull/2074))
 
+- Added `close_range(2)` on Linux
+
 ## [0.27.1] - 2023-08-28
 
 ### Fixed